コードをちょっと分割
